reorganization step 1
