+++ codes/surrogates/surrogates.py
@@ -10,7 +10,7 @@ from torch import Tensor, nn
 from torch.utils.data import DataLoader
 from tqdm import tqdm
 
-from utils import create_model_dir
+from codes.utils import create_model_dir
 
 
 class AbstractSurrogateModel(ABC, nn.Module):
