+++ data/osu2008/data.hdf5
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:b676735cb73f32312696886c3d0190dffc7b5fbbf230a953095c3ee8aed814f5
+size 23208992
