+++ pytest.ini
@@ -1,0 +1,5 @@
+[pytest]
+testpaths = "test"
+pythonpath = "."
+python_functions = "test_*"
+python_classes = "Test*"
